Refactor V1.0
